Increase a lot the logging when someone asks for access to a git repo
The current information logged when someone asks/tries to access a
git repo via the aclchecker script is very low. With this commit
we will track in much finer details who asks for what and what the
answer was in the web server's log since aclchecker calls the
internal API endpoints and basically follows whatever decision that
endpoint makes.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>