Fix typo and button label on the bugzilla assignee feature
Fix a typo in the title of the button used to edit/change the
bugzilla assignee overrides.
Change as well the label of the button from Update to Edit as
the former seemed to imply to some people that it would go and
fetch that information from somewhere else while it is meant
to offer a way to edit the information.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>