Stop setting a backref on relations that are viewonly
From sqlalachemy 1.3.14 relations that are specifying a backref when
viewonly=True will issue a warning and this will be disallowed in
future releases of sqlalchemy.
So, after checking that these backref were not used in the project,
we have simply removed them.

Source: https://github.com/sqlalchemy/sqlalchemy/releases/tag/rel_1_3_14
which also explains why this behavior is not desired.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>