Rework the detailed provided for the issue stats
Basically, we want to know the trend of open tickets, so we take
the current number of tickets and for each week we remove from this
number the number of opened tickets and add the closed tickets (which
were thus still open at the beginning of the week).

We also only want to ignore closed tickets without a closed_at date
if they are in the window of tickets we're looking at (a year). So
if the most recent ticket without a closed_at date is older than a
year, then we assume that all the tickets have a closed_at date and
move on.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>