Improve loading tickets from git
If a ticket we're loading as a close_status that isn't defined at
the project level, create it.
If the ticket we're loading as a status that isn't Open or Closed
consider it Closed and use that status as close_status.
Finally, record the closed_at value from the ticket that is being
loaded if there is such value.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>