Upgrade the Docker-based dev environment
The containers now all inherit from a common Fedora 31 image, with only
Python 3 packages installed. This saves disk space, and creates a
consistent testing environment.

The base Dockerfile installs Python packages from the distro repository
instead of PyPi. This creates a stable installation that isn't dependent
on the latest version of any package, so it should work with less
adjustment in the future.