Use symlink instead of pip3 install which isn't always working as desired
Basically pip3 install doesn't like being run as root but if it's not
run as root it doesn't quite work because it runs into permission error.
So instead, let's just have a symlink that will make pagure available in
the python path.

We're also creating a symlink to ensure /usr/bin/python exists (and it
will be python3) since hooks are relying on it.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>