If pagure is set up for local auth, allow git push via https using it
If pagure supports local authentication, let the users provide their
username and password when doing a git push via http.
If the username and password provided didn't match what is in the DB,
fallback to check if the user isn't using an API token as for the
other authentication systems.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>