Make the tests PagureFlaskApiIssuetests.test_api_assign_issue more robust
It was relying on a hard-coded ACL id which is based on the order in
which the ACL were loaded into the database. As new ACL are added
this order may change (and do!) resulting in a weird failing test.
With this change, we're directly retrieving the ACL id using the
place this ACL has in the configuration file, making this much more
robust.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>