Add the new git auth backend to the default list
The get_git_auth_helper method in pagure.lib.git_auth has logic
to rely on the default/provided git auth if for some reasons it
fails to load them from the pkg_resources.
This method did not account for the new git auth backend that was
recently added and this commit fixes this.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>