Add PAGURE_PLUGINS_CONFIG setting in pagure configuration file.
Plugins are imported from a separate configuration file that is loaded either
by setting a PAGURE_PLUGIN environment variable, or with the --plugins flag of
the runserver.py script.
This commit introduces a new variable called PAGURE_PLUGINS_CONFIG that
replaces PAGURE_PLUGIN. The new variable preserves the behavior of the old
variable, and additionally allows to specify the plugins configuration file
directly from the pagure main configuration.
A new "Plugins" section has also been added to the pagure documentation for
documenting the new changes.