Make the stats page use the new stats API endpoint
We originally designed this code to use the existing API endpoint
with an argument to say that it wants the detailed version of the
stats. We later changed our mind and move the details to a new
API endpoint, but apparently we forgot to update the UI to point
to the new API endpoint (while all the JS code is compatible/expected
this data).

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>