Use WhiteNoise to serve static assets for the Pagure web interface
In order to make it easier to run Pagure in a containerized environment,
the web frontend needs to be capable of serving all of the frontend static
files. This change introduces WhiteNoise as a dependency that would enable
the Flask application itself to serve the static assets correctly.

This makes it possible to run a Pagure server without Apache or Nginx at all
in a containerized environment for all of the main functions.

Note that the releases folder is still not served through WhiteNoise, and
deployments that have uploading releases enabled will need something to
serve them.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>