Introducing the boards feature
This feature introduce the presentation of tickets in some sort of
light weight kanban board.
When creating a board, it is associated with a tag of the project.
Adding ticket to this tag will automatically add the ticket to the
board (and vice-versa, un-tagging will remove it).
The boards can have any number of status (columns) which are defined
by the user but one of them must be marked as the default status to
which tickets are added when they are tagged.
Some status can be marked as "Close" in which case they will close
the ticket when it reaches that status. If a "Close status" is
specified, that close status will be applied when closing the ticket.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>