Allow having a dedicated loggin configuration for the git hooks
Pagure's logging can now be configured in ``LOGGING`` as before
for both the web application as well as the git hooks.
But if needed/desired, the git hooks can have a different logging
configuration that will be specified in ``LOGGING_GIT_HOOKS``.
If ``LOGGING_GIT_HOOKS`` is not specified (default) the ``LOGGING``
configuration will be used instead.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>