Add a NOGITHOOKS configuration key
This configuration key allows to entirely skip the git hooks.
This is useful for the test suite as it allows to block some
action from happening upon git push, which is done in a separate
process. The action can then be triggered in the main process
where they can be checked and thus tested.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>