When checking if messages were sent via a rebase, do not run the git hooks
Otherwise, the PR is updated in a separate process that is inaccessible
from pytest. That process is then the one sending the fedora-messaging
notifications and suddenly when we check the UI, none of the expected
notifications were sent.

So instead, we entirely skip the git hooks. The PR gets updated when
the rebase finishes, in the process run by pytest and we are able to
check the notifications sent.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>