Rework how PR flags are show in the UI
Now that the PR flag API actually flag commit rather than PR objects,
we presentation of these flags in the UI is changed as well.
We both include the flags in the list of commits, as we do on the
commit list and in addition, we show the flags of the last commit
on the PR just above the main comment area.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>