Allow people with a fork to have a working drop-down for opening new PRs
Basically, the drop-down for opening new PR from the main page of a
main project was conditioned on having commit right on that main
project. However, if you have a fork and no access to the main
project, the drop-down was not returning.
Now, we're checking if you have access to the main project or if
you have a fork and in either case the JS ensuring the drop-down
will populate fine.

Fixes https://pagure.io/fedora-infrastructure/issue/9497

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>