Allow to specify a default branch for all projects hosted on an instance
Basically, with this feature we can define the default git branch for all
the new projects created in that pagure instance.
This default branch can always be overriden by the user upon creation, but
if the user does not specify a default branch, it will be using this one.

Fixes https://pagure.io/pagure/issue/5054

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>