Fix the pagination links in the "My PRs" page
We were using the macro but not importing it so when it was actually try
to add the pagination, it would raise a 500 error.
Similarly, we were not passing to the template the "page" variable so we
couldn't pass it to the macro and if we tried we would get another 500
error.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>