Port pagure to a F33-based vagrant image for development
This commit basically moves the Vagrantfile example file
from Fedora 30 to Fedora 33, adjusts the python path from
python 3.7 to python 3.9 and change the order of the argument
to the celery CLI for the services.
This last one seems new compared to previous celery releases but
without it the services are no longer starting.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>