Allow updating the target branch when editing a PR
With this commit we will be able to update a PR to change its target
branch in cases that branch disappears.
For example, you have a PR opened against the "master" branch but the
project decides to change its default branch to "main" and delete the
"master" branch to avoid confusion.
Now, the PR that was opened against "master" will error saying it cannot
find the target branch for the PR (and the diff of the PR will thus be
empty). With this commit, one will be able to edit the PR and change
the target branch from "master" to "main", thus allowing the review to
continue.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>