Be consistent in the messages send
The JSON representation of the User object has an 'emails' field
not an 'email' and that field is not present in the public JSON
representation of the User objects (thus not in notifications sent
on the message bus). We can thus drop entirely the 'email' field
that was present until now.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>