Add support for noggin with the FAS auth backend
The particularity of this backend is to require that users sign the
Fedora Project Contributor Agreement. With the new Auth system in
Fedora, the contributors who have signed this agreement are in a
dedicated group while it was separated before. That does mean that
pagure needs to ask for the members of this group upon login,
otherwise it has no way to know if the user is a member of that group
(ie, if the user signed the agreement).

This commit fixes that.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>