Reorder celery arguments (and s/info/INFO/)
After upgrading my server to Debian bullseye, I noticed that some of
the pagure systemd services were not properly starting anymore.
celery was complaining that it didn't recognize the "-A" argument in
the "celery worker" command.

A quick web search told me that the "-A" argument must be passed right
after the "celery" command.  After doing that, celery was now
complaining that it didn't recognize "info" as a proper loglevel.
After reading the error message, I noticed that it did have an
INFO (all caps) loglevel, so after uppercasing the word celery finally
started.

I'm not entirely sure when celery started being more pedantic about
its arguments (the same systemd unit was working with buster's
celery), but I think it's a good idea to reflect the needed changes
upstream so that others are not bit by this issue.

Signed-off-by: Sergio Durigan Junior <sergiodj@sergiodj.net>