srcfpo theme: left-align the lines in description
The description texts are taken from package %descriptions, and our packaging
guidelines specify that the text is 80-column wrapped. Centering the lines of
the text leads to poorly-looking christmas-tree-like layout. People occasionally
use lists and such, and this looks especially bad if the vertical alignment
is not preserved.

I think it's nice to leave the paragraph centered, but with individual lines
left-aligned within the paragraph:

  Blah blah  blah  blah  blah  blah  blah  blah  blah blah blah blah
  blah  blah  blah
  blah  blah  blah  blah  blah  blah  blah  blah  blah blah blah:
  - foo
  - bar
  blah  blah  blah  blah  blah  blah  blah  blah  blah  blah  blah

Tested with firefox-92.0.1-1.fc35.x86_64, chromium-93.0.4577.63-1.fc35.x86_64.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>