Warn users when a PR contains some characters
Unicode bi-directional characters can be present but unseen and thus
missed during the review.
With this PR, we create a list of characters that we want to warn the
users about if present in a PR. Since that list is configurable, it
can be extended as needed/desired.

This relates to the CVE: CVE-2021-42574 and CVE-2021-42694
More can be found about it at:
https://access.redhat.com/security/vulnerabilities/RHSB-2021-007

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>