Fix object of "rebased onto" comment
When you rebase a pull request, you get an automatic comment
that says "rebased onto <commit id>". However, the commit ID is
always the first commit of the pull request, not the commit ID
it was actually rebased onto. e.g. in
https://pagure.io/fedora-qa/relval/pull-request/21
right now the most recent message is "rebased onto 0fe88a2", and
the first commit in the PR series is
"Cleanup unreachble code for coverage reasons.", 0fe88a2. I think
this should make it correct.

Signed-off-by: Adam Williamson <awilliam@redhat.com>