Update pagure CI guide
I noticed that the notification from post build is not sent to PR and after some
investigation I found out that the credentials need to contain CI token instead
of API key. Unfortunately you can't look at the secret text in Jenkins
configuration, so it took some time till I was able to figure this out.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>