fix(theme): long words in source nav break layout
The 'srcfpo' theme adds additional information to the commits side navigation, which might contain long words and then break the layout.
In that case the content appeared below the navigation and not side-by-side with the commits overview or commit details.
To fix this, a text break will be enforced in the relevant div of the 'srcfpo' theme.
Also the column width in global 'commit.html' aligned with other template files by setting to 'col-2'.

Fixes: https://pagure.io/pagure/issue/5353