fix: grammar of "log in"
"Login" is a noun or adjective. The verb is "log in". This becomes obvious when
one attempts present continuous: "I'm logging in", and not "I'm loginning".

I also adjusted grammar in surrounding context when I noticed something obvious.

I had a few fesco tickets open today without being logged in and the wrong form
was hurting me a little every time ;)