feat(run-tests-container): use local repo and branch by default
default repo changed from 'https://pagure.io/pagure.git' to '/wrkdir' and
default branch changed from 'master' to 'wrkdirbranch', logic added
that will use the git repo and active branch in the current directory to perform tests.
It's not necessary to push changes to a remote repository to run all or a specific test case.

A repo name without 'http://' or 'https://' will be interpreted as local path and handled accordingly.

Both can be overridden to a URL and branch name for the same behavior as before.

Also some cleanup / refactoring to reduce redundant code.

Signed-off-by: Dominik Wombacher <dominik@wombacher.cc>