feat(run-tests-container): move logging outside container
Writing logs to '/results/' inside the container was done via the entrypoint script.
To have a full log coverage, also from build and shell operations, it's now
part of the run-tests-container.py3 script and writes to dev/results_<test-container-name>.
Every log file has the current unix timestamp as prefix, files are not overwritten.

Signed-off-by: Dominik Wombacher <dominik@wombacher.cc>