tests: Drop mock 'pagure.lib.git._maybe_wait'
An unknown Async / wait / timing issue caused the following test to fail:
    FAILED
    tests/test_pagure_flask_dump_load_ticket.py::PagureFlaskDumpLoadTicketTests::test_dumping_reloading_ticket
    - Exception: Unable to find object

This came from the override of 'pagure.lib.git._maybe_wait', default behavior is 'pass', mock made it a 'result.get()'.
Likely that the version bump to a newer Celery major release was the root cause.
Without this enforced wait, the test passes.